Write the q_new, q_free, q_insert, q_remove
In q_new, we need to allocate the memory for the list_head and init it. If fall, return NULL.
In q_free, checking the head whether exist and going through the every nodes to free them and their value.
In q_insert, using the strcpy to copy the string to value. Using the function of list.h to add node to list.
In q_remove, using list_first_entry to find the first entry of list. First, we need to copy the value to sp. Need to notice the size of value and sp.