Write q_reverse and q_reverseK and q_swap
In q_reverseK, Setting a buffer of list_head to save the list that were reversed. Using a pointer to find the nodes of the list k.
In q_swap, go through the each node of list. Avoid to move the node of head.
In q_reverse, using 3 pointer to do the reverse.